+++ packages/language-core/src/codegen/script/template.ts
@@ -42,7 +42,7 @@ function* generateTemplateCtx(options: ScriptCodegenOptions): Generator<Code> {
     } else {
       yield `const __VLS_defineExposeUnrefs = __VLS_defineExpose as any as UnwrapRefs<typeof __VLS_defineExpose>${newLine}`
     }
-    yield `const __MPX_ctx = { ...__MPX_dollars, ...__VLS_defineProps, ...__VLS_defineExposeUnrefs }${endOfLine}`
+    yield `const __MPX_ctx = { ...__VLS_defineProps, ...__VLS_defineExposeUnrefs }${endOfLine}`
   } else {
     yield `const __MPX_ctx = __VLS_defineComponent${endOfLine}`
   }