+++ packages/language-core/src/utils/parseJsonUsingComponents.ts
@@ -1,0 +1,50 @@
+import type { SfcJsonBlockUsingComponents } from '../types'
+import type * as ts from 'typescript'
+
+export function parseUsingComponents(
+  ts: typeof import('typescript'),
+  sourceFile: ts.SourceFile,
+): SfcJsonBlockUsingComponents {
+  const usingComponents: SfcJsonBlockUsingComponents = new Map()
+
+  if (
+    !sourceFile.statements.length ||
+    !ts.isExpressionStatement(sourceFile.statements[0])
+  ) {
+    return usingComponents
+  }
+
+  const objectLiteral = sourceFile.statements[0].expression
+  if (!ts.isObjectLiteralExpression(objectLiteral)) {
+    return usingComponents
+  }
+
+  const usingComponentsProperty = objectLiteral.properties.find(
+    (prop): prop is ts.PropertyAssignment =>
+      ts.isPropertyAssignment(prop) &&
+      ts.isStringLiteral(prop.name) &&
+      prop.name.text === 'usingComponents',
+  )
+
+  if (
+    usingComponentsProperty &&
+    ts.isObjectLiteralExpression(usingComponentsProperty.initializer)
+  ) {
+    for (const prop of usingComponentsProperty.initializer.properties) {
+      if (
+        ts.isPropertyAssignment(prop) &&
+        ts.isStringLiteral(prop.name) &&
+        ts.isStringLiteral(prop.initializer)
+      ) {
+        // eg: name: 'list', value: '../components/list'
+        usingComponents.set(prop.name.text, {
+          text: prop.initializer.text,
+          offset: prop.initializer.getStart(sourceFile),
+          nameOffset: prop.name.getStart(sourceFile),
+        })
+      }
+    }
+  }
+
+  return usingComponents
+}