+++ packages/language-service/src/plugins/mpx-sfc-json-json.ts
@@ -1,5 +1,8 @@
+import type * as vscode from 'vscode-languageserver-protocol'
 import type { LanguageServicePlugin } from '@volar/language-service'
 import { create as baseCreate } from 'volar-service-json'
+import { URI } from 'vscode-uri'
+import { MpxVirtualCode } from '@mpxjs/language-core'
 
 export function create(): LanguageServicePlugin {
   const base = baseCreate({
@@ -18,5 +21,77 @@ export function create(): LanguageServicePlugin {
     },
   })
 
-  return base
+  return {
+    name: 'mpx-json-json',
+
+    capabilities: {
+      ...base.capabilities,
+      documentLinkProvider: {},
+    },
+
+    create(context) {
+      if (!context.project.mpx) {
+        return {}
+      }
+
+      const baseInstance = base.create(context)
+
+      return {
+        ...baseInstance,
+
+        async provideDocumentLinks(document) {
+          const uri = URI.parse(document.uri)
+          const decoded = context.decodeEmbeddedDocumentUri(uri)
+          if (!decoded) {
+            return
+          }
+          const [documentUri, embeddedCodeId] = decoded
+          if (embeddedCodeId !== 'json_json') {
+            return
+          }
+          const sourceScript = context.language.scripts.get(documentUri)
+          const virtualCode =
+            sourceScript?.generated?.embeddedCodes.get(embeddedCodeId)
+          if (!sourceScript?.generated || virtualCode?.id !== 'json_json') {
+            return
+          }
+
+          const root = sourceScript.generated.root
+          if (!(root instanceof MpxVirtualCode) || !root.sfc.json) {
+            return
+          }
+
+          const result: vscode.DocumentLink[] = []
+
+          const usingComponents = await root.sfc.json.resolveUsingComponents
+
+          if (!usingComponents?.size) {
+            return result
+          }
+
+          for (const [
+            _,
+            {
+              text: componentPath,
+              offset: componentPathOffset,
+              realFilename: targetFilePath,
+            },
+          ] of usingComponents) {
+            result.push({
+              range: {
+                start: document.positionAt(componentPathOffset),
+                end: document.positionAt(
+                  componentPathOffset + componentPath.length,
+                ),
+              },
+              target: targetFilePath,
+              tooltip: `自定义组件：${componentPath}`,
+            })
+          }
+
+          return result
+        },
+      }
+    },
+  }
 }