+++ packages/language-service/src/plugins/mpx-sfc-template-links.ts
@@ -5,32 +5,37 @@ import { MpxVirtualCode, Sfc, tsCodegen } from '@mpxjs/language-core'
 
 export function create(): LanguageServicePlugin {
   return {
-    name: 'mpx-document-links',
+    name: 'mpx-template-links',
+
     capabilities: {
       documentLinkProvider: {},
     },
+
     create(context) {
       return {
-        provideDocumentLinks(document) {
+        async provideDocumentLinks(document) {
           const uri = URI.parse(document.uri)
           const decoded = context.decodeEmbeddedDocumentUri(uri)
-          const sourceScript =
-            decoded && context.language.scripts.get(decoded[0])
+          if (!decoded) {
+            return
+          }
+          const [documentUri, embeddedCodeId] = decoded
+          const sourceScript = context.language.scripts.get(documentUri)
           const virtualCode =
-            decoded && sourceScript?.generated?.embeddedCodes.get(decoded[1])
+            sourceScript?.generated?.embeddedCodes.get(embeddedCodeId)
           if (!sourceScript?.generated || virtualCode?.id !== 'template') {
             return
           }
-
           const root = sourceScript.generated.root
           if (!(root instanceof MpxVirtualCode)) {
             return
           }
 
           const result: vscode.DocumentLink[] = []
-
           const { sfc } = root
           const codegen = tsCodegen.get(sfc)
+
+          // #region document link for style class
           const scopedClasses =
             codegen?.getGeneratedTemplate()?.scopedClasses ?? []
           const styleClasses = new Map<
@@ -99,6 +104,39 @@ export function create(): LanguageServicePlugin {
               }
             }
           }
+          // #endregion
+
+          // #region document link for component tag
+          const templateNodeTags =
+            codegen?.getGeneratedTemplate()?.templateNodeTags ?? []
+          const usingComponents = await sfc.json?.resolveUsingComponents
+          if (usingComponents?.size) {
+            for (const nodeTag of templateNodeTags) {
+              const {
+                name: componentTag,
+                startTagOffset,
+                endTagOffset,
+              } = nodeTag
+              if (!usingComponents.has(componentTag) || !startTagOffset) {
+                continue
+              }
+              const { text: componentPath, realFilename: targetFilePath } =
+                usingComponents.get(componentTag)!
+              const addLink = (offset: number) => {
+                result.push({
+                  range: {
+                    start: document.positionAt(offset),
+                    end: document.positionAt(offset + componentTag.length),
+                  },
+                  target: targetFilePath,
+                  tooltip: `自定义组件：${componentPath}`,
+                })
+              }
+              addLink(startTagOffset)
+              endTagOffset && addLink(endTagOffset)
+            }
+          }
+          // #endregion
 
           return result
         },