+++ packages/language-service/src/plugins/mpx-sfc.ts
@@ -58,7 +58,9 @@ export function create(): LanguageServicePlugin {
 
   return {
     ...htmlService,
+
     name: 'mpx-sfc',
+
     capabilities: {
       ...htmlService.capabilities,
       diagnosticProvider: {
@@ -66,6 +68,7 @@ export function create(): LanguageServicePlugin {
         workspaceDiagnostics: false,
       },
     },
+
     create(context) {
       const htmlServiceInstance = htmlService.create(context)
 
@@ -130,11 +133,13 @@ export function create(): LanguageServicePlugin {
             const { startTagEnd = Infinity, endTagStart = -Infinity } =
               template ?? {}
 
-            for (const error of mpxSfc.errors) {
+            for (const error of mpxSfc.errors.concat(
+              sfc.template?.errors ?? [],
+            )) {
               if ('code' in error) {
                 const start = error.loc?.start.offset ?? 0
                 const end = error.loc?.end.offset ?? 0
-                if (end < startTagEnd || start >= endTagStart) {
+                if (start > startTagEnd || end <= endTagStart) {
                   sfcErrors.push({
                     range: {
                       start: document.positionAt(start),